+++ Source/rmpq/common.h
@@ -116,10 +116,10 @@ bool validateHash(void const* buf, uint32_t size, void const* expected);
 bool validateHash(File file, void const* expected);
 
 #define ID_WAVE         0x46464952
-bool pkzip_compress(void* in, size_t in_size, void* out, size_t* out_size);
-bool pkzip_decompress(void* in, size_t in_size, void* out, size_t* out_size);
-//bool multi_compress(void* in, size_t in_size, void* out, size_t* out_size, uint8_t mask, void* temp);
-bool multi_decompress(void* in, size_t in_size, void* out, size_t* out_size, void* temp = nullptr);
+bool pkzip_compress(void* in, uint32_t in_size, void* out, uint32_t* out_size);
+bool pkzip_decompress(void* in, uint32_t in_size, void* out, uint32_t* out_size);
+//bool multi_compress(void* in, uint32_t in_size, void* out, uint32_t* out_size, uint8_t mask, void* temp);
+bool multi_decompress(void* in, uint32_t in_size, void* out, uint32_t* out_size, void* temp = nullptr);
 
 const char* path_name(const char* path);
 const char* path_ext(const char* path);
